this should bring us closer to py2/3 compat
